changed naming for scheduled exercise interface
